Fix Todo List icons
