--- src/App/Exercises/ReactExercises/LocalDevAndFetch/EditIcon/EditIcon.css
@@ -1,12 +1,0 @@
-.bin-icon path {
-  stroke: #000;
-}
-
-.bin-icon:hover path {
-  stroke: #06a0f9;
-  cursor: pointer;
-}
-
-.bin-icon--error path {
-  stroke: #ff0000;
-}
