--- src/App/Exercises/ReactExercises/LocalDevAndFetch/EditIcon/EditIcon.jsx
@@ -1,69 +1,0 @@
-import './EditIcon.css';
-
-export function EditIcon() {
-  return (
-    <svg
-      width="100"
-      height="26"
-      viewBox="0 0 100 26"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M10.7314 25.2638C10.5553 25.2641 10.3808 25.2245 10.218 25.1475C10.0553 25.0704 9.90741 24.9574 9.783 24.8148L0.392971 14.0842C0.268399 13.9419 0.169579 13.7729 0.102153 13.5869C0.0347267 13.4009 1.51875e-05 13.2016 4.98265e-09 13.0002C-1.51775e-05 12.7989 0.0346664 12.5996 0.102064 12.4136C0.169462 12.2276 0.268257 12.0585 0.392807 11.9162C0.517357 11.7738 0.665224 11.6609 0.827965 11.5838C0.990706 11.5068 1.16513 11.4671 1.34129 11.4671C1.51745 11.467 1.69188 11.5067 1.85463 11.5837C2.01739 11.6607 2.16527 11.7736 2.28984 11.916L10.7314 21.5627L28.5629 1.18461C28.8145 0.897439 29.1556 0.736197 29.5112 0.736328C29.8669 0.73646 30.2079 0.897955 30.4593 1.18532C30.7108 1.47268 30.8521 1.86239 30.8523 2.2688C30.8524 2.67521 30.7114 3.06505 30.4601 3.35265L11.68 24.8148C11.5556 24.9574 11.4077 25.0704 11.2449 25.1474C11.0821 25.2245 10.9076 25.264 10.7314 25.2638Z"
-        fill="#23213D"
-      />
-      <path
-        d="M54.7407 3.99394L43.4489 16.0345C43.0225 16.4917 42.6099 17.3923 42.5274 18.0158L42.0185 22.5051C41.8397 24.1262 42.995 25.2346 44.5904 24.9575L49.0192 24.1955C49.6381 24.0846 50.5046 23.6274 50.9309 23.1563L62.2227 11.1157C64.1758 9.03739 65.056 6.66808 62.0164 3.77225C58.9906 0.904136 56.6937 1.9156 54.7407 3.99394Z"
-        stroke="#23213D"
-        strokeWidth="1.9375"
-        strokeMiterlimit="10"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M53.0002 6C53.5525 9.75146 56.4296 12.6194 60.0002 13"
-        stroke="#23213D"
-        strokeWidth="1.9375"
-        strokeMiterlimit="10"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M96.2632 6.47831C92.6557 6.12081 89.0265 5.93665 85.4082 5.93665C83.2632 5.93665 81.1182 6.04498 78.9732 6.26165L76.7632 6.47831"
-        stroke="#E12B1B"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M82.7214 5.38413L82.9598 3.96496C83.1331 2.93579 83.2631 2.16663 85.0939 2.16663H87.9323C89.7631 2.16663 89.9039 2.97913 90.0664 3.97579L90.3048 5.38413"
-        stroke="#E12B1B"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M93.934 9.90173L93.2298 20.8109C93.1106 22.5117 93.0131 23.8334 89.9906 23.8334H83.0356C80.0131 23.8334 79.9156 22.5117 79.7965 20.8109L79.0923 9.90173"
-        stroke="#E12B1B"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M84.7041 17.875H88.3116"
-        stroke="#E12B1B"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M83.8049 13.5416H89.2216"
-        stroke="#E12B1B"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-}
