+++ src/App/Images/Icons/CheckMarkIcon.jsx
@@ -1,0 +1,19 @@
+import './Icon.css';
+
+export function CheckMarkIcon() {
+  return (
+    <svg
+      className="icon--check-mark"
+      width="31"
+      height="26"
+      viewBox="0 0 31 26"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M10.7314 25.2638C10.5553 25.2641 10.3808 25.2245 10.218 25.1475C10.0553 25.0704 9.90741 24.9574 9.783 24.8148L0.392971 14.0842C0.268399 13.9419 0.169579 13.7729 0.102153 13.5869C0.0347267 13.4009 1.51875e-05 13.2016 4.98265e-09 13.0002C-1.51775e-05 12.7989 0.0346664 12.5996 0.102064 12.4136C0.169462 12.2276 0.268257 12.0585 0.392807 11.9162C0.517357 11.7738 0.665224 11.6609 0.827965 11.5838C0.990706 11.5068 1.16513 11.4671 1.34129 11.4671C1.51745 11.467 1.69188 11.5067 1.85463 11.5837C2.01739 11.6607 2.16527 11.7736 2.28984 11.916L10.7314 21.5627L28.5629 1.18461C28.8145 0.897439 29.1556 0.736197 29.5112 0.736328C29.8669 0.73646 30.2079 0.897955 30.4593 1.18532C30.7108 1.47268 30.8521 1.86239 30.8523 2.2688C30.8524 2.67521 30.7114 3.06505 30.4601 3.35265L11.68 24.8148C11.5556 24.9574 11.4077 25.0704 11.2449 25.1474C11.0821 25.2245 10.9076 25.264 10.7314 25.2638Z"
+        fill="#23213D"
+      />
+    </svg>
+  );
+}
