+++ src/App/Images/Icons/EditIcon.jsx
@@ -1,0 +1,31 @@
+import './Icon.css';
+
+export function EditIcon() {
+  return (
+    <svg
+      className="icon"
+      width="24"
+      height="25"
+      viewBox="0 0 24 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M13.7407 2.99394L2.44888 15.0345C2.02252 15.4917 1.60991 16.3923 1.52738 17.0158L1.01849 21.5051C0.839696 23.1262 1.99501 24.2346 3.59044 23.9575L8.01915 23.1955C8.63807 23.0846 9.50455 22.6274 9.93092 22.1563L21.2227 10.1157C23.1758 8.03739 24.056 5.66808 21.0164 2.77225C17.9906 -0.0958643 15.6937 0.915597 13.7407 2.99394Z"
+        stroke="#23213D"
+        strokeWidth="1.9375"
+        strokeMiterlimit="10"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M12.0002 5C12.5525 8.75146 15.4296 11.6194 19.0002 12"
+        stroke="#23213D"
+        strokeWidth="1.9375"
+        strokeMiterlimit="10"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
