+++ src/App/Images/Icons/Icon.css
@@ -1,0 +1,19 @@
+.icon path {
+  stroke: #000;
+}
+
+.icon--check-mark path {
+  fill: #000;
+}
+
+.icon:hover path {
+  stroke: #06a0f9;
+}
+
+.icon--check-mark:hover path {
+  fill: #06a0f9;
+}
+
+.icon--error path {
+  stroke: #ff0000;
+}
